03_npz_numpy_pipeline/requirements.txt

numpy
pillow
scikit-learn
matplotlib
psutil
tensorflow==2.*
